updated optimized: bidirectional
